refactor: extract into `ActivitiesHelper`
