+++ app/helpers/application_helper.rb
@@ -40,19 +40,6 @@ module ApplicationHelper
     link_to label, url, class: 'pdf-link'
   end
 
-  def activities_link_for_trackable trackable
-    activities_link trackable.class.name, trackable.id
-  end
-
-  def activities_link trackable_type, trackable_id
-    return unless trackable_type
-    return if trackable_type.in? Activity::DEPRECARD_TRACKABLE_TYPES
-
-    type = trackable_type.constantize.base_class
-    url = activities_path(trackable_type: type, trackable_id: trackable_id)
-    link_to (antcat_icon("filter") + 'Activities'), url, class: "btn-normal filter-activities-link"
-  end
-
   def yes_no_options_for_select value
     options_for_select([["Yes", "true"], ["No", "false"]], value)
   end
