+++ app/helpers/taxon_helper.rb
@@ -1,13 +1,4 @@
 module TaxonHelper
-  def taxon_link_or_deleted_string id, deleted_label = nil
-    taxon = Taxon.find_by(id: id)
-    if taxon
-      taxon.link_to_taxon
-    else
-      deleted_label || "##{id} [deleted]"
-    end
-  end
-
   def taxon_change_history taxon
     return if taxon.old?
     change = taxon.last_change
